INT-1768: Remove old indexes tab
