+++ src/app/home/collection.jade
@@ -15,13 +15,10 @@
           a Explain Plan
         li(role='presentation', data-hook='index-tab', id='index-tab')
           a Indexes
-        li(role='presentation', data-hook='indexes-new-tab', id='indexes-new-tab')
-          a Indexes New
     .row
       div(data-hook='refine-bar-subview')
 
   div(data-hook='document-subview')
   div(data-hook='schema-subview')
   div(data-hook='explain-subview')
-  div(data-hook='index-subview')
-  div(data-hook='indexes-new-subview' class='index-container')
+  div(data-hook='index-subview' class='index-container')
