+++ src/app/index.less
@@ -14,7 +14,6 @@
 @import "sidebar/index.less";
 @import "network-optin/index.less";
 @import "identify/index.less";
-@import "indexes/index.less";
 @import "explain-plan/index.less";
 
 @import "../auto-update/index.less";
@@ -28,3 +27,4 @@
 @import "../internal-packages/status/styles/index.less";
 @import "../internal-packages/query/styles/index.less";
 @import "../internal-packages/schema/styles/index.less";
+@import "../internal-packages/indexes/styles/index.less";
