--- src/app/indexes/index-item.jade
@@ -1,46 +1,0 @@
-tr
-  td.name-column
-    div.name= model.name
-    div(data-hook='index-definition-subview')
-  td.type-column
-    - var type = model.type
-    if type
-      if type === 'text'
-        // tooltips required on text indexes
-        div(class="property #{type}", data-toggle='tooltip', title=text_tooltip)
-          =type
-          i.link
-      else
-        //- no tooltips on other types
-        div(class="property #{type}")
-           = type
-           i.link
-  td.size-column
-    div.quantity(data-hook='index-size')
-    div.unit(data-hook='index-size-unit')
-    div.progress(data-toggle='tooltip', title='#{model.relativeSize}% compared to largest index')
-      div.progress-bar(role='progressbar', data-hook='progressbar')
-  td.usage-column
-    span.usage(data-toggle='tooltip')
-      div.quantity(data-hook='usage-count')
-      div.usage-since since&nbsp;
-        span(data-hook='usage-since')
-  td.property-column
-    div.properties
-      each prop in model.properties
-        if prop === 'ttl'
-          div(class='property', data-toggle='tooltip', title=ttl_tooltip)
-            = prop
-            i.link
-        else if prop === 'partial'
-          div(class='property', data-toggle='tooltip', title=partial_tooltip)
-            = prop
-            i.link
-        else
-          div(class='property')
-            = prop
-            i.link
-      if model.cardinality === 'compound'
-        div.property.cardinality
-          = model.cardinality
-          i.link
