--- src/app/indexes/index.jade
@@ -1,23 +1,0 @@
-.index-container
-  .flexbox-fix
-  .column-container
-    .column.main
-      table
-        thead
-          tr
-            th.name(data-hook='th-name') Name and Definition
-              i.sort.fa.fa-fw
-              i.fa.fa-fw.fa-sort
-            th(data-hook='th-type') Type
-              i.sort.fa.fa-fw
-              i.fa.fa-fw.fa-sort
-            th(data-hook='th-size') Size
-              i.sort.fa.fa-fw
-              i.fa.fa-fw.fa-sort
-            th.usage-column(data-hook='th-usage') Usage
-              i.sort.fa.fa-fw
-              i.fa.fa-fw.fa-sort
-            th(data-hook='th-properties') Properties
-              i.sort.fa.fa-fw
-              i.fa.fa-fw.fa-sort
-        tbody.table(data-hook='indexes')
