+++ src/internal-packages/indexes/index.js
@@ -2,6 +2,7 @@
 
 const app = require('ampersand-app');
 const Indexes = require('./lib/component/indexes');
+const NameColumn = require('./lib/component/name-column');
 const Action = require('./lib/action/index-actions');
 
 /**
@@ -9,6 +10,7 @@ const Action = require('./lib/action/index-actions');
  */
 function activate() {
   app.appRegistry.registerComponent('Component::Indexes::Indexes', Indexes);
+  app.appRegistry.registerComponent('Component::Indexes::NameColumn', NameColumn);
   app.appRegistry.registerAction('Action::Indexes::LoadIndexes', Action.loadIndexes);
 }
 
@@ -17,6 +19,7 @@ function activate() {
  */
 function deactivate() {
   app.appRegistry.deregisterComponent('Component::Indexes::Indexes');
+  app.appRegistry.deregisterComponent('Component::Indexes::NameColumn');
   app.appRegistry.deregisterAction('Action::Indexes::LoadIndexes');
 }
